+++ Android/app/src/main/java/com/example/rocketcorner/LoginActivity.java
@@ -1,0 +1,121 @@
+package com.example.rocketcorner;
+
+//import android.content.Intent;
+import android.content.Context;
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Patterns;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.ProgressBar;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+
+public class LoginActivity extends AppCompatActivity implements View.OnClickListener {
+    private TextView register;
+    private EditText editTextEmail, editTextPassword;
+    private Button signIn;
+
+    private FirebaseAuth mAuth;
+    private ProgressBar progressBar;
+    private Button button;
+    public static final String BASE_URL = "http://rocketcorner.herokuapp.com/";
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        // testing
+        Intent intent = MainActivity.getIntent(getApplicationContext());
+        startActivity(intent);
+
+        setContentView(R.layout.activity_main);
+
+        register = (TextView) findViewById(R.id.register);
+        register.setOnClickListener(this);
+
+        signIn = (Button) findViewById(R.id.signIn);
+        signIn.setOnClickListener(this);
+
+        editTextEmail = (EditText) findViewById(R.id.email);
+        editTextPassword = (EditText) findViewById(R.id.password);
+
+        progressBar = (ProgressBar) findViewById(R.id.progressBar);
+
+        
+    }
+
+
+    public static Intent getIntent(Context context){
+        Intent intent = new Intent(context, LoginActivity.class);
+        return intent;
+    }
+
+    @Override
+    public void onClick(View v) {
+        switch (v.getId()){
+            case R.id.register:
+                startActivity(new Intent(this, RegisterUser.class));
+                break;
+
+            case R.id.signIn:
+                userLogin();
+                break;
+        }
+    }
+
+
+    private void userLogin(){
+        String email = editTextEmail.getText().toString().trim();
+        String password = editTextPassword.getText().toString().trim();
+
+        if(email.isEmpty()){
+            editTextEmail.setError("Email is required");
+            editTextEmail.requestFocus();
+            return;
+        }
+        if(!Patterns.EMAIL_ADDRESS.matcher(email).matches()){
+            editTextEmail.setError("Please enter a valid email");
+            editTextEmail.requestFocus();
+            return;
+
+        }
+        if(password.isEmpty()){
+            editTextPassword.setError("Password is required");
+            editTextPassword.requestFocus();
+            return;
+        }
+        if(password.length() < 6){
+            editTextPassword.setError("Min password length is 6 characters");
+            editTextPassword.requestFocus();
+            return;
+        }
+
+
+        progressBar.setVisibility(View.VISIBLE);
+        mAuth = FirebaseAuth.getInstance();
+
+        mAuth.signInWithEmailAndPassword(email, password).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+            @Override
+            public void onComplete(@NonNull Task<AuthResult> task) {
+
+                if(task.isSuccessful()){
+                    Intent intent = MainActivity.getIntent(getApplicationContext());
+                    startActivity(intent);
+                }else{
+                    Toast.makeText(LoginActivity.this, "Failed to login", Toast.LENGTH_LONG).show();
+                }
+            }
+        });
+
+    }
+
+}
