+++ Android/app/src/main/java/com/example/rocketcorner/MainActivity.java
@@ -1,117 +1,60 @@
 package com.example.rocketcorner;
 
-//import android.content.Intent;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
+
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Patterns;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ProgressBar;
-import android.widget.TextView;
-import android.widget.Toast;
+import android.view.MenuItem;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
+import com.example.rocketcorner.fragments.AccountFragment;
+import com.example.rocketcorner.fragments.FundsFragment;
+import com.example.rocketcorner.fragments.ShopFragment;
+import com.google.android.material.bottomnavigation.BottomNavigationView;
+import com.google.android.material.navigation.NavigationBarView;
 
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-
-public class MainActivity extends AppCompatActivity implements View.OnClickListener {
-    private TextView register;
-    private EditText editTextEmail, editTextPassword;
-    private Button signIn;
-
-    private FirebaseAuth mAuth;
-    private ProgressBar progressBar;
-    private Button button;
-    public static final String BASE_URL = "http://rocketcorner.herokuapp.com/";
+public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_profile);
 
-        register = (TextView) findViewById(R.id.register);
-        register.setOnClickListener(this);
+        BottomNavigationView bottomNav = findViewById(R.id.bottom_nav);
 
-        signIn = (Button) findViewById(R.id.signIn);
-        signIn.setOnClickListener(this);
 
-        editTextEmail = (EditText) findViewById(R.id.email);
-        editTextPassword = (EditText) findViewById(R.id.password);
+        bottomNav.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
+            @Override
+            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
+                Fragment selectedFragment = null;
 
-        progressBar = (ProgressBar) findViewById(R.id.progressBar);
+                switch(item.getItemId()){
+                    case R.id.ic_account:
+                        selectedFragment = new AccountFragment();
+                        break;
+                    case R.id.ic_shop:
+                        selectedFragment = new ShopFragment();
+                        break;
+                    case R.id.ic_money:
+                        selectedFragment = new FundsFragment();
+                        break;
+                }
 
-        
+                getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, selectedFragment).commit();
+                return true;
+            }
+        });
+
+        if(savedInstanceState == null){
+            getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, new AccountFragment()).commit();
+        }
     }
-
 
     public static Intent getIntent(Context context){
         Intent intent = new Intent(context, MainActivity.class);
         return intent;
-    }
-
-    @Override
-    public void onClick(View v) {
-        switch (v.getId()){
-            case R.id.register:
-                startActivity(new Intent(this, RegisterUser.class));
-                break;
-
-            case R.id.signIn:
-                userLogin();
-                break;
-        }
-    }
-
-
-    private void userLogin(){
-        String email = editTextEmail.getText().toString().trim();
-        String password = editTextPassword.getText().toString().trim();
-
-        if(email.isEmpty()){
-            editTextEmail.setError("Email is required");
-            editTextEmail.requestFocus();
-            return;
-        }
-        if(!Patterns.EMAIL_ADDRESS.matcher(email).matches()){
-            editTextEmail.setError("Please enter a valid email");
-            editTextEmail.requestFocus();
-            return;
-
-        }
-        if(password.isEmpty()){
-            editTextPassword.setError("Password is required");
-            editTextPassword.requestFocus();
-            return;
-        }
-        if(password.length() < 6){
-            editTextPassword.setError("Min password length is 6 characters");
-            editTextPassword.requestFocus();
-            return;
-        }
-
-
-        progressBar.setVisibility(View.VISIBLE);
-        mAuth = FirebaseAuth.getInstance();
-
-        mAuth.signInWithEmailAndPassword(email, password).addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-            @Override
-            public void onComplete(@NonNull Task<AuthResult> task) {
-
-                if(task.isSuccessful()){
-                    Intent intent = profileActivity.getIntent(getApplicationContext());
-                    startActivity(intent);
-                }else{
-                    Toast.makeText(MainActivity.this, "Failed to login", Toast.LENGTH_LONG).show();
-                }
-            }
-        });
-
     }
 
 }
