+++ Android/app/src/main/java/com/example/rocketcorner/RegisterUser.java
@@ -52,7 +52,7 @@ public class RegisterUser extends AppCompatActivity implements  View.OnClickList
     public void onClick(View v) {
         switch (v.getId()){
             case R.id.banner:
-                startActivity(new Intent(this, MainActivity.class));
+                startActivity(new Intent(this, LoginActivity.class));
                 break;
             case R.id.registerUser:
                 registerUser();
