--- Android/app/src/main/java/com/example/rocketcorner/profileActivity.java
@@ -1,60 +1,0 @@
-package com.example.rocketcorner;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
-
-import android.content.Context;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.MenuItem;
-
-import com.example.rocketcorner.fragments.AccountFragment;
-import com.example.rocketcorner.fragments.FundsFragment;
-import com.example.rocketcorner.fragments.ShopFragment;
-import com.google.android.material.bottomnavigation.BottomNavigationView;
-import com.google.android.material.navigation.NavigationBarView;
-
-public class profileActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile);
-
-        BottomNavigationView bottomNav = findViewById(R.id.bottom_nav);
-
-
-        bottomNav.setOnItemSelectedListener(new NavigationBarView.OnItemSelectedListener() {
-            @Override
-            public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-                Fragment selectedFragment = null;
-
-                switch(item.getItemId()){
-                    case R.id.ic_account:
-                        selectedFragment = new AccountFragment();
-                        break;
-                    case R.id.ic_shop:
-                        selectedFragment = new ShopFragment();
-                        break;
-                    case R.id.ic_money:
-                        selectedFragment = new FundsFragment();
-                        break;
-                }
-
-                getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, selectedFragment).commit();
-                return true;
-            }
-        });
-
-        if(savedInstanceState == null){
-            getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, new AccountFragment()).commit();
-        }
-    }
-
-    public static Intent getIntent(Context context){
-        Intent intent = new Intent(context, profileActivity.class);
-        return intent;
-    }
-
-}
